[JENKINS-42861] - Properly deprecate the HudsonExceptionNote implementation (#2811)
* [JENKINS-42861] - Properly deprecate the HudsonExceptionNote implementation

It is a replacement of the original fix in https://github.com/jenkinsci/jenkins/pull/2808.

The implementation provides deprecates the annotation of the new stacktraces, but it retains the binary and the persisted data compatibility, which were missing in the original PR.
In the longer-term the hyperlinks should be replaced by the best possible equivalent (JIRA search, grepcode or whatever).

* [JENKINS-42861] - Cleanup the deprecated functionality

(cherry picked from commit 77a9f026f88d7cbbceca12f8f5645277a6aadaa0)