Removing extra slash in static resources urls
The way `RESOURCE_PATH` is used is e.g. `Jenkins.getInstance().getRootUrl() + Functions.getResourcePath()`  while under `getRootUrl` there is a `return Util.ensureEndsWith(url,"/");` which means the combination becomes `http://root//static/bla`. This change will make it `http://root/static/bla`. The double slash generally works fine except for a bug in a specific reverse proxy (traefik) which fails to load the resource because of the double slashes. See https://github.com/containous/traefik/issues/1323

I realize the path `/static/` means to convey that it should be at the root and not in a subdirectory - but it seems there is a requirement to call `getRootUrl` anyway. I did not do the due diligence to verify this fix does not break anything else so I hope there's a test that runs on this PR.