[JENKINS-43496] - Add handling of the null Node#createComputer() result.
it is a follow-up to https://github.com/jenkinsci/jenkins/pull/2836#discussion_r110604865

De-facto many Cloud plugins return `null` in `Node#createLauncher()`, but it has never been documented.
In order to prevent possible API misusages in the future, I have added annotations and fixed handling of the extension point in `AbstractCIBase#updateComputer()` which may fail in the case of `null` or `RuntimeException` in the Node implementation.