[JENKINS-48821] - Don't forget estimated duration of asynchronous jobs (#3329)
* Don't forget estimated duration of asynchronous jobs.

When jobs are executed, the estimated duration is calculated, memorized
and reseted when the job is finished.

But when the job is executed asynchronously (e.g. as for pipeline
scripts), the execution is just triggered. Therefore the estimated
duration should remain - and not be reseted after triggering.

JENKINS-48821: Time remaining for all pipeline jobs is N/A
JENKINS-49616: Estimated time remaining NA

* Reset cached estimated duration also when asynchronous job ends.

(cherry picked from commit 3fc392108ae382a4616964aa8bbb6b8cb9dff0dc)