[JENKINS-50969] core site build is broken, needed for core taglibs reference
The special old module for doxia seems not needed.
I suspect that it is because since 2012 when it was released, the Markdown
support got done in standard.

IIUC this was pulling in some very old doxia versions, which was
incompatible with maven-fluido-skin which I needed to bump from current
1.3.0 to 1.7 for https://github.com/jenkins-infra/core-taglibs-report-generator
to work again.