[JENKINS-50412] Fingerprints logs should be less verbose (#3373)
* JENKINS-50412 - Fingerprints logs should be less verbose

Eliminate the logging line that says
possibly trimming /var/jenkins_home/fingerprints/[...]
This line fills up the fingerprints log file with repetitive, useless information.
The operation it calls provides its own logging that can be turned on if detailed
logging is actually needed instead of running this one repeatedly.

Also, while in the area, convert a couple of FileFilter inner classes to lambdas
and inline them. This reduces unnecessary boilerplate and improves clarity.

In order to make this change testable, I elected to remove the final restriction on the
class and annotate the class as restricted. This allows me to create unit tests for
this change and also verify existing capability at the appropriate level.

* Watch out for varying line endings.

Use a reliable check.

* Correct a few minor issues caught in review.

* Change to correct access modifier.