Fix sortable.js to handle "Never" and null dates
Build Failure Analyzer plugin can put "Never" in the "Last Seen" column,
which broke matching/sorting in sortable.js. This change handles
"Never" (so it includes it as a Date) and non-matching strings in
a column determined to be date, and treats them as equal to the Epoch,
so they sort as very old.