[JENKINS-51965] bundle JAXB
After a failed initial attempt, falling back to a simpler option.

Servlet containers and EE containers are designed to tolerate this kind of setup where
the base system brings one version of JAXB and the app brings another
version because it's a common practice. When run in servlet containers
in Java 9+, this will ensure JAXB is available.

The version of JAXB bundled works with Java8.