JENKINS-52159: Queue.maintain - tolerate misbehaving sorter.sortBuildableItems()
Any plugin that implements the sort step could throw an exception which would break
the maintain() step. As sorting is really an optional step, it really should not be fatal.

Instead, we wrap the step in a try block, log it, and move on with life.

(cherry picked from commit 1083a3f5c648e87464e25815779690eed1fb3296)