cli: Fix up usage message
The properties file naively assumed that leading spaces would be honored.
-- They are not. To get some indentation, one needs to escape a space first.

Further, for no particular reason, the last paragraph had a hard wrap.

Finally, since this commit has to reindent the entire block, fix the alignment of the `:`s
for most of the options.
(-auth is a mess and should be rewritten)