JENKINS-21336: ADMINISTER should not imply RUN_SCRIPTS
This introduces a new feature flag, hudson.security.Permission.administerImpliesRunScripts which is set to true by default. Setting this to false changes the default permission behavior to use RUN_SCRIPTS as the root level permission.

This also introduces a separated permission registry so that plugins may eagerly register permission instances on startup to help alleviate problems such as in JENKINS-17200.

Also relates to internal issue CORE-1068.