Java 11 test failure: no need to log at WARNING.
java.lang.AssertionError: expected:<[]> but was:<[Running Jenkins with Java 11 which is available in the preview mode only. A custom experimental update center will be used: https://updates.jenkins.io/temporary-experimental-java11/]>
	at org.junit.Assert.fail(Assert.java:88)
	at org.junit.Assert.failNotEquals(Assert.java:834)
	at org.junit.Assert.assertEquals(Assert.java:118)
	at org.junit.Assert.assertEquals(Assert.java:144)
	at jenkins.model.StartupTest.noWarnings(StartupTest.java:45)

Note that this was marked as a flake because Surefire reruns the test
(apparently in the same JVM despite reuse.forks=false)
and the message is printed in a static initializer, which gets run just once.