Use a fixed version of ATH
We should not use master, but explicitly bump regularly to latest state.
That gives a reproducible state to PRs and master.

In case something goes wrong on ATH, this avoids spending time wondering
if something changed upstream or not, or the change is the actual
thing responsible for failure.

Bottom line: change one thing at a time to go faster.

Here I'm choosing arbitrarily as one last merge commit a few days ago when
we know Core PRs were NOT failing because of `runATH`.

The intent, if this gets a green build, is to then file PRs regularly to
stay as close as possible to ATH master branch state.