Update core to support user-creation notifications
[JENKINS-55305] Add user creation listener method

Four new methods added to the SecurityListener class
to enable support for listening to Jenkins user-account
creation events.

[JENKINS-55305] Update SecurityListener class

[JENKINS-55307] Add new private-realm unit tests

Added unit tests to exercise the updated methods
of the HudsonPrivateSecurityRealm class.

[JENKINS-55307] Update realm user-creation methods

[JENKINS-55305] Remove usage of listener method

Updated the SecurityListener and HudsonPrivateSecurityRealm
classes, removing usage of the SecurityListener failedToCreateUser()
method, as no valid Jenkins use cases could be identified for the
method at this point in time.