Problem: [JENKINS-55512] Jenkins stalls in quietDown with complex jobs
Solution: Allow to run tasks that are spawned by tasks,
and only block other build causes if the Jenkins server
is shutting down or otherwise quiescing. Otherwise, it
waits indefinitely for the parent job to complete while
its subsequent children (which the job blocks waiting
for) never run.

Added a couple of "FIXME" comments for possible future
refinement that is not in direct scope of fixing this
deadlock.

Signed-off-by: Jim Klimov <EvgenyKlimov@eaton.com>