JENKINS-55834 Typo in hudsonHomeIsFull, because of escaping apostrophes (#3898)
* Update index.properties

* Update index.properties

Whenever you are using MessageFormat you should be aware that the single quote character (') fulfils a special purpose inside message patterns. The single quote is used to represent a section within the message pattern that will not be formatted. A single quote itself must be escaped by using two single quotes ('').

* Update index.properties

Changed "Can't" to can not

* Update core/src/main/resources/hudson/diagnosis/HudsonHomeDiskUsageMonitor/index.properties

Co-Authored-By: Francois-D <46973576+Francois-D@users.noreply.github.com>