Remove incorrect annotation that overrides return type of AbstractItem#getParentGroup
This annotation overrode the return type of AbstractItem#getParentGroup
to return Jenkins instead of ItemGroup. This would not compile, when a
normal ItemGroup was returned, as that does not extend Jenkins.

We ran into this issue when using Mockito with the new ByteBuddy mockmaker
that would throw a WrongTypeOfReturnValue as it expected a Jenkins
value, but saw an ItemGroup instead.