Add f:secretTextarea UI analogous to f:password
Just as f:password is provided for secrets instead of f:textinput,
this f:secretTextarea is provided for secrets instead of f:textarea
where said secrets are multiline in nature. While this component does
not provide any way to view an existing secret, it provides UIs to
add and replace secrets.

Signed-off-by: Matt Sicker <boards@gmail.com>