Removed BZIP2 support and added distinction between the two snappy formats
Following the recommendations, this commit drops the BZIP2 format, since
it seems not sensible for log files, given the newer compression formats.

Also, this commit makes the distinction between the framed and raw snappy
formats, that are usually distinguished by their extension.

The framed format also has a magic number prefix, but for that the first few
bytes of the file or stream would need to be read.