Removed dependency on deprecated getLogFile() method
The previous code was missing a change to getLogFile(), that was
not applied due to the deprecation of getLogFile() in the upstream code.

This led to the new files never being read. The new code will pick up
compressed log files regardless of their extension, as long as they are
named "log" or "log.*" -- with the uncompressed version always being
preferred.

Also:
 - Extended the logging to give some additional feedback to the users
   in case of errors or missing files
 - Rephrased the API docs to no longer refer to log files
   (See: https://github.com/jenkinsci/jenkins/pull/3724)