[JENKINS-55945] IllegalStateException trying to get user. (#4107)
This change catches the exception and returns false, invalidating the user session. This behavior is sensible, because if we are unable to validate the user we must assume the session is invalid.
Catching the error at this point reportedly helps with some scenarios and shouldn't cause any further harm.

(cherry picked from commit 8349cebb61a9aa73511f243465a6c5f57cfadd2f)