[JENKINS-58993] do not allow save() to be called before we have loaded the config.
if save() is called by a plugin (or core) before we have loaded we
will
not have a correctly populated in memory model in which to save, which
would then be persisted to disk which will result in unacceptable
dataloss.

This is obviously not a good thing.
Because this is such a severe issue the error state is logged in core as
well as denying the save with an IOException.