[JENKINS-58993] we need to set configLoaded earlier
As changing the builds and workspaces property will cause Jenkins to
save its configuration immediatley after it was loaded we need to do
this before we potentially save the config after it was just loaded.

This makes sense anyway as XStream has finished its deserialization so
out in menory model is correct.