[JENKINS-53322] Remove patched dom4j fork (#4130)
* [WIP] [JENKINS-58170] Reference an incremental stapler version.

One that uses the standard dom4j library instead of our fork.
We want to run this through CI and testing frameworks.

* Tweak a test assertion to be more precise.

Change the assertion to be more precise about what it is looking for. Make it check for the element's text and exclude the text of any children.

* Use released stapler version

Stapler has now been released with the changes to eliminate the dom4j fork.