Make pipeline container scrollable
Hi,

I've added a hint how to do pipeline container scrollable and more intuitive to use. Currently it becomes too wide when I use many columns and page design becomes broken.

I'm not familiar with the actual template engine and ap logic, but see how it's entered and fixed it in my browser. Now I ask you to make the final change in the code to make this naive error disappear for all of us.

Thank you