[JENKINS-60913] Remove network discovery services  (#4460)
* Remove multicast discovery services.

* Refine clean up with DNSMultiCast.

Doesn't need to be referenced in the other pom any more.
We need to keep the class and the property around for a while.
Jenkins Test Harness sets the property. Until we can get all
versions aligned so it isn't needed, it's best to keep the property
doing nothing.

* Deprecate remaining class fragment.

* Remove unneeded dependency.

* Improve code comments for external usages.

Co-Authored-By: Jesse Glick <jglick@cloudbees.com>

* Remove multicast discovery services.

* Refine clean up with DNSMultiCast.

Doesn't need to be referenced in the other pom any more.
We need to keep the class and the property around for a while.
Jenkins Test Harness sets the property. Until we can get all
versions aligned so it isn't needed, it's best to keep the property
doing nothing.

* Deprecate remaining class fragment.

* Remove unneeded dependency.

* Add log warning messages.

* Don't need to implement the interface anymore.

* Fix initializer.

Change capitalization for logger.

* Add URLs to log messages.

Co-Authored-By: Daniel Beck <1831569+daniel-beck@users.noreply.github.com>

Co-authored-by: Jesse Glick <jglick@cloudbees.com>
Co-authored-by: Daniel Beck <1831569+daniel-beck@users.noreply.github.com>