[JENKINS-43889] Environment from pre-checkout SimpleBuildWrapper not always teared down
An Environment contributed by a pre-checkout SimpleBuildWrapper will not be teared
down is the build stops too early (for instance during SCM checkout).
Although JENKINS-43889 was not originally about SimpleBuildWrapper, but about a
"regular" BuildWrapper (in ssh-agent-plugin), it really is the same issue (with the
Environment being added by the preCheckout method).