[JENKINS-43889] AbstractBuildExecution: ensure Environments are always teared down
Before this change, in an AbstractBuildExecution, tearing down Environments used to
be left to the doRun() implementations.  But this comes too late to actually tear down
the ones which have been setup before checkout (by BuildWrapper#preCheckout, or
RunListener#setUpEnvironment).

Tearing down Environments is now done directly in AbstractBuildExecution#run(), in a
finally block.

The BuildExecution#doRun implementation is modified to take that into account.
But because in plugins (for instance maven-plugin) there are still implementations
of doRun() which could tear down Environments themselves, we use a dummy
Environment (TearDownCheckEnvironment) as a marker to check whether teardown
has been done already, and avoid doing it twice.