[JENKINS-61224] Different colors for monitor count in old vs new UI
Both sets of colors provide good contrast. It now sets a "stronger" red background with white text for the old UI (makes it more noticeable in the black navbar) and a softer red with a darker text for the new UI (looks better with the blue bg). Both come from boostrap styles ("badge-danger" vs "alert-danger")