[JENKINS-61224] Removing the "monitor[s] text next to the bell (#4555)
* [JENKINS-61224] Removed span tag including the "monitor[s]" text

* [JENKINS-61224] Removed monitors from the i18n properties files

* [JENKINS-61224] Different colors for monitor count in old vs new UI

Both sets of colors provide good contrast. It now sets a "stronger" red background with white text for the old UI (makes it more noticeable in the black navbar) and a softer red with a darker text for the new UI (looks better with the blue bg). Both come from boostrap styles ("badge-danger" vs "alert-danger")

* [JENKINS-61224] Trivial commit on comment to force re-build

* [JENKINS-61224] Adding the "-v2" suffix to new UI vars