[JENKINS-29144] Add perform() taking EnvVars to SimpleBuildStep
Provides a default implementation for both overloads:
- if the old method is called, it forwards to the new method, passing
  the environment specified for the Run
  - this means that when a new implementer (who has overridden the new
    overload) gets called by an old client, they get the run
    environment only (fine for freestyle, only the initial environment
    in a pipeline)
- if the new method is called, it forwards to the old method, ignoring
  the passed environment
  - in this case, the old method MUST be overridden
  - this handles the case of an old implementer getting called by a
    new client
    - no change in functionality; implementer still gets the same
      environment to work with (via the Run)

BuildStepCompatibilityLayer has _not_ been changed yet to call the new
overload, mainly because it would just get the environment to pass the
same way the default implementation in SimpleBuildStep does (i.e.
passing the result of calling getEnvironment() on the Run).

Implementers of SimpleBuildStep (like ArtifactArchiver and
Fingerprinter) similarly have _not_ been changed. This is mainly
because they get Run's environment to apply substitution, and that
should not be done using an environment passed in from a pipeline
step. So if they switch to the new signature, their code should also
be adjusted to only expand environment variables when the Run is an
AbstractBuild.

With this change, a PR can be made to workflow-basic-steps-plugin,
to extend CoreStep to pass in the environment variables from the
step context. The combination of both PRs should then fix
JENKINS-29144 and allow the creation of Builders that are
automatically available as pipeline steps with proper honoring of
things like withEnv and global tools.