[JENKINS-29144] Update ArtifactArchiver and Fingerprinter to use the new SimpleBuildStep API
Both will only use the environment for substitution when the Run being passed is an AbstractBuild, following the guidelines that for pipelines, the user does their own substitution in Groovy.
This causes a small behaviour change; before, even in pipelines, substitution would be done by these steps using the (tiny) initial environment.