[JENKINS-46175] SimpleBuildWrappers without workspace
This adds predicates to SimpleBuildWrapper and
SimpleBuildWrapper.Disposer, to indicate whether a workspace and/or
launcher are required. As with SimpleBuildStep these all default
to true, to preserve the old behaviour by default.

This also adds explicit nullability annotations to all affected
methods, assuming non-null for the context, build and listener.