[JENKINS-62688] use id in User.toString()
Change toString() to use the id rather than fullName.

This is because a full name may not be unique and as such when used in log
messages we get something that is less than helpful.

as such we should print the user's Id so we can actually identify the
user.