Add a checkItem() to the "old" LabelValidator.
By default, this calls the regular check() if the item is an AbstractProject;
otherwise, it returns OK.
This allows a plugin to implement an override for that method in order to have its
validator applied to Jobs too, without needing to bump their Jenkins dependency in
order to get the new LabelValidator.

As a result, LabelExpression.validate() now always uses the "old" LabelValidators,
calling this new method.