[JENKINS - 63332] Radio button's move around when clicking them
When the user clicked the radio button, there's a foldable block would show. But when the block was empty, it would show an empty tr with 2px padding. I try to set 'visibility' to 'hidden' and 'position' to 'absolute' to hide the foldable block that relative to the radio button that was not clicked by user.