removed dc-license.txt
the license was supposed to be for the old original Hudson icons.
these have been long since replaced with pngs, svgs and new artwork so
should no longer be required.  (it was also in the root and not along
with other licences in about).  Given this is no longer needed as the
icons are long gone we can remove this file.