[JENKINS-64465] Fix race condition on HudsonFilter#reset
If `Jenkins#setSecurityRealm` is called concurrently from multiple threads there is a
potential race condition on the authentication filters registration.

`Jenkins#setSecurityRealm` calls `filter.reset()`, `HudsonFilter#reset` is not
synchronized however it is replacing the old filter with the new filter (if there was one)
 which is not a thread safe operation.

The effect of this race condition is the wrong auth filter being registered.