JENKINS-1962: Support rendering local time based on browser's formatting
The earlier implementation relied on the server timezone. While this works
for colocated teams, this is not great for distributed teams.

This change relies on passing milliseconds since epoch as an attribute to a
HTML tag, and then using javascript on page load to set the body of the tag
to a correctly formatted timestamp.