[JENKINS-36707] Add `CI` environment variable
Introduces a standard `CI` env var that is always set to "true" in order
to allow pipelines and tools to know they're executing in a continuous
integration context. This has emerged as a best practice used by most
other popular CI solutions, and a number of tools look for this variable
in order to make CI-appropriate optimizations (cf. the ticket).

Closes JENKINS-36707.