[JENKINS-66379] upgrade winstone to pickup upstream fix
The change uses the JVM default keystore.  This allows jenkins to be
used for TLS termination even when JKS is not the default keystore, or
is not allowed for other restrictions.

If someone has a non default keystore yet is using a keystore here then
this will break on upgrade, however it is unlikely that they would be
having a non default keystore as the normal reason for doing that is due
to a restriction like FIPS-140 and given the keystore we require has a
key using JDK here would be non compliant.