[JENKINS-66379] upgrade winstone to allow non JKS keystores for https certificates (#5670)
* [JENKINS-66379] upgrade winstone to pickup upstream fix

The change uses the JVM default keystore.  This allows jenkins to be
used for TLS termination even when JKS is not the default keystore, or
is not allowed for other restrictions.

If someone has a non default keystore yet is using a keystore here then
this will break on upgrade, however it is unlikely that they would be
having a non default keystore as the normal reason for doing that is due
to a restriction like FIPS-140 and given the keystore we require has a
key using JDK here would be non compliant.

* Update war/pom.xml

update winstone to release version (5.21)

(cherry picked from commit d3acf6ac9dd529326a6f11745a1b6ccc16e1a94c)