Fix bug in TextparameterDefinition (#5704)
TextParameterDefinition did not override getDefaultParameterValue()
method and instead relied on the implementation in
StringParameterDefinition. In most cases this wasn't a problem, but in
some cases, for example when triggering build through
buildWithParameters API without specifying the parameter, then the
parameter object would be of the wrong type (StringParameterValue as
opposed to TextParameterValue). This causes problems with visualization
in parameters view as well as incorrect values when rebuild plugin is
used.

Change-Id: If4e11bdafcd94c9c83462983f384fa80ae2c4d9c
(cherry picked from commit cfec42e28b802bde699ef13b82f2f45206bcb306)