[JENKINS-67075] Strip a possible daemon-injected prefix when restarting the server.
At least the Debian packages rely on the `daemon` command to start the
server with appropriate parameters, leveraging the /etc/init.d/jenkins
shell script and the variables set in the /etc/default/jenkins config
file.

The init script calls `/usr/bin/daemon --name=$NAME […]`, with NAME
defaulting to `jenkins`. Starting with daemon 0.7.0-1 (as found in
Debian 11 “Bullseye” and in Ubuntu since Hirsute Hippo), using the
`--name` option results in the specified value's being prepended to
the program being started.

This can be checked with:

    root@jenkins:~# pgrep -f jenkins: -a
    73109 jenkins: /usr/bin/java -Djava.awt.headless=true -jar /usr/share/jenkins/jenkins.war --webroot=/var/cache/jenkins/war --httpPort=8080

    root@jenkins:~# tr '\0' '\n' < /proc/73109/cmdline
    jenkins: /usr/bin/java
    -Djava.awt.headless=true
    -jar
    /usr/share/jenkins/jenkins.war
    --webroot=/var/cache/jenkins/war
    --httpPort=8080

Therefore, detect and remove such a prefix from the first argument
(the executable). Since the daemon's name is configurable, don't
hardcode "jenkins: ", but strip everything until the first slash if
it's preceeded by a colon and a space.

Signed-off-by: Cyril Brulebois <cyril@debamax.com>