Prevent one redirect in AsynchPeople
When clicking on a link from the "People" view, as the links do not include a trailing slash, they are redirected by Stapler to the same url with trailing slash. This could be avoided by just adding the trailing slash when displaying the URL.

Voluntarily not changing `getUrl()` to not break expected behavior.