Revert "[JENKINS-67237] `BuildTrigger` waits until the dependency graph has been updated (#6101)"
https://issues.jenkins.io/browse/JENKINS-68189 reports that downstream
jobs are no longer executed in Jenkins 2.341.

https://issues.jenkins.io/browse/JENKINS-67237 reports that Downstream
jobs are not guaranteed to build when the job configs have been updated
recently.

Comments in https://github.com/jenkinsci/jenkins/pull/6101 recommend
that this pull request be reverted while a fix is being developed.
This change is the revert of the pull request.

The revert was generated by pressing the "revert" button on the GitHub
interface, then using the resulting commit in this pull request.

This reverts commit 292382fe05f0e1c0188113d2f909144452a5011f.