[JENKINS-43727] show assigned labels of agents
If enabled this monitor will show the labels of agents on the /computer
overview page.
The maximum number of labels to show is configurable. When more labels
are defined, those will be shown when hovering over the list.

This is no using model-link as this doesn't work together with the
hovering.