clear interrupt flag before latch.await in trigger shutdown
A latent interrupt flag set before we call `latch.await` will immediately throw an error when we call the await function.
It is better to discard this interrupt flag if it was set before we start `latch.await`. 

This way the await can still be interrupted, but only if the interrupt is called on its thread _after_ the await actually began.