[JENKINS-68591] Use new layout for remote class loader stats
The groovy generated layout is not as attractive as the jelly layout,
but this seems like an improvement over the current layout.  If others
have a better approach to improve it, I am open to suggestions.

The table expands to full width of the page.  That makes the table more
difficult to read than the old table, because the data values are placed
at the far right of the screen and the data label is near the far left
of the screen.