Fix tests that where failing due to incorrect assumptions
now that we actually have the plugins as expected in JenkinsRule 4 tests
where failing for a variety of reasons

1. loading a detached plugin would downgrade the version requiring a
   restart and failing the test
2. the test was checking that nothing was installed

installing detached plugins has now been skipped where the plugin being
installed was already available due to being a maven dependency.  This
also means we only have onve version of the plugin on the classpath -
prior to this both the classes from the maven dependency and the classes
from the plugin would have been on the uber classpath creating
interesting results.

tests have been updated where there are changes between the detached
version of the plugin and the plugin from the maven dependency.

Where the test was making invalid assumptions this has also been fixed.